Added metrics to JobLogger, and made JobLogger configurable.
This commit adds GC and deserialization to the data
output by the job logger.  It also adds metrics about
how much shuffle data was fetched locally and how much time
the local fetch took. Finally, it makes the job logger
configurable, so it can be enabled (but is disabled by
default) with configuration parameters.